+++ back/src/main/java/com/curso/spring/rest/models/dao/UsuarioDao.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UsuarioDao extends JpaRepository<Usuario, Long> {
+public interface UsuarioDao extends JpaRepository<Usuario, Integer> {
 
     Usuario findByUsername(String username);
